moving card file writing and card file reading into core pkg so we can leverage this logic in the git realm
