+++ cards/files-realm/tracker.ts
@@ -1,11 +1,8 @@
 import { IndexingOperations } from '@cardstack/core/indexer';
-import { readdirSync, statSync, readJsonSync, readFileSync } from 'fs-extra';
 import { join } from 'path';
-import { assertSingleResourceDoc } from '@cardstack/hub/jsonapi';
-import merge from 'lodash/merge';
-import { Card } from '@cardstack/core/card';
 import logger from '@cardstack/logger';
 import { UpstreamDocument, UpstreamIdentity } from '@cardstack/core/document';
+import { Entry, readCard, crawl } from '@cardstack/core/card-file-utils';
 
 const log = logger('files-realm-tracker');
 
@@ -18,7 +15,7 @@ export class FilesTracker {
   private ids: Map<string, UpstreamIdentity> = new Map();
 
   async update(directory: string, ops: IndexingOperations): Promise<void> {
-    let now = this.crawl(directory);
+    let now = crawl(directory);
     let previous = this.state.get(directory);
 
     if (!previous) {
@@ -36,7 +33,7 @@ export class FilesTracker {
 
         try {
           let cardDir = join(directory, name);
-          let json = this.assembleCard(cardDir, entry);
+          let json = readCard(cardDir, entry);
           let upstreamId = {
             csId: json.data.attributes!.csId as string,
             csOriginalRealm: json.data.attributes!.csOriginalRealm as string,
@@ -85,83 +82,6 @@ export class FilesTracker {
   }
 
   notifyFileDidChange(_fullPath: string): void {}
-
-  private crawl(cardsDirectory: string): Map<string, Entry> {
-    let output: Map<string, Entry> = new Map();
-    for (let name of readdirSync(cardsDirectory)) {
-      if (name.startsWith('.') || name === 'node_modules') {
-        continue;
-      }
-      let fullName = join(cardsDirectory, name);
-      let stat = statSync(fullName);
-      if (stat.isDirectory()) {
-        output.set(name, this.crawl(fullName));
-      } else {
-        output.set(name, { mtime: stat.mtime.getDate(), size: stat.size });
-      }
-    }
-    return output;
-  }
-
-  private assembleCard(cardDirectory: string, files: Map<string, Entry>) {
-    let pkg;
-    try {
-      pkg = readJsonSync(join(cardDirectory, 'package.json'));
-    } catch (err) {
-      if (err.code === 'ENOENT') {
-        throw new Error(`Card does not have a valid package.json file`);
-      }
-      throw err;
-    }
-
-    let json;
-    try {
-      json = readJsonSync(join(cardDirectory, 'card.json'));
-      assertSingleResourceDoc(json);
-    } catch (err) {
-      if (err.code === 'ENOENT') {
-        throw new Error(`Card does not have a valid card.json file`);
-      }
-      if ('isCardstackError' in err) {
-        throw new Error(`card.json is invalid because: ${err}`);
-      }
-      throw err;
-    }
-
-    // ensure we have an attributes object
-    merge(json, {
-      data: {
-        attributes: {},
-        meta: {
-          cardDir: cardDirectory,
-        },
-      },
-    });
-
-    // then ensure that csFiles reflects our true on disk files only
-    json.data.attributes!.csFiles = this.loadFiles(cardDirectory, files, ['package.json', 'card.json']);
-
-    // and our peerDeps match the ones from package.json
-    // @ts-ignore
-    json.data.attributes!.csPeerDependencies = pkg.peerDependencies;
-    return json;
-  }
-
-  private loadFiles(dir: string, files: Map<string, Entry>, exclude: string[] = []) {
-    let output: NonNullable<Card['csFiles']> = Object.create(null);
-    for (let [name, entry] of files) {
-      if (exclude.includes(name)) {
-        continue;
-      }
-      let fullName = join(dir, name);
-      if (entry instanceof Map) {
-        output[name] = this.loadFiles(fullName, entry);
-      } else {
-        output[name] = readFileSync(fullName, 'utf8');
-      }
-    }
-    return output;
-  }
 }
 
 function changed(previous: Entry | undefined, current: Map<string, Entry>): boolean {
@@ -199,8 +119,6 @@ function changed(previous: Entry | undefined, current: Map<string, Entry>): bool
   }
   return false;
 }
-
-type Entry = { mtime: number; size: number } | Map<string, Entry>;
 
 declare module '@cardstack/hub/dependency-injection' {
   interface KnownServices {
