+++ packages/core/package.json
@@ -8,6 +8,7 @@
   "dependencies": {
     "@cardstack/logger": "^0.2.1",
     "assert-never": "^1.2.0",
+    "fs-extra": "^8.1.0",
     "http-status-codes": "^1.4.0",
     "json-typescript": "^1.1.2",
     "jsonapi-typescript": "^0.1.3",
